Guess The Number v 0.1
Q:
Write a program where the computer randomly generates a number between a range. The user needs to guess what the number is. If the user guesses wrong, tell them their guess is either too high, or too low. This will get you started with the random library if you haven't already used it.